requirements.txt
flask
openai
python-dotenv
PyPDF2
numpy
